+++ src/main/java/com/ats/tril/model/PoDetail.java
@@ -11,11 +11,11 @@ public class PoDetail {
 	private String itemCode;
 	private String itemDesc; 
 	private String itemUom; 
-	private int itemQty; 
+	private float itemQty; 
 	private float itemRate; 
-	private int mrnQty; 
-	private int pendingQty; 
-	private int indedQty; 
+	private float mrnQty; 
+	private float pendingQty; 
+	private float indedQty; 
 	private float discPer; 
 	private float discValue; 
 	private int schDays; 
@@ -30,7 +30,7 @@ public class PoDetail {
 	private float freightValue; 
 	private float otherChargesAfter; 
 	private float landingCost;
-	private int balanceQty;
+	private float balanceQty;
 	private String indMNo;
 	
 	public String getIndMNo() {
@@ -87,10 +87,10 @@ public class PoDetail {
 	public void setItemUom(String itemUom) {
 		this.itemUom = itemUom;
 	}
-	public int getItemQty() {
+	public float getItemQty() {
 		return itemQty;
 	}
-	public void setItemQty(int itemQty) {
+	public void setItemQty(float itemQty) {
 		this.itemQty = itemQty;
 	}
 	public float getItemRate() {
@@ -99,22 +99,22 @@ public class PoDetail {
 	public void setItemRate(float itemRate) {
 		this.itemRate = itemRate;
 	}
-	public int getMrnQty() {
+	public float getMrnQty() {
 		return mrnQty;
 	}
-	public void setMrnQty(int mrnQty) {
+	public void setMrnQty(float mrnQty) {
 		this.mrnQty = mrnQty;
 	}
-	public int getPendingQty() {
+	public float getPendingQty() {
 		return pendingQty;
 	}
-	public void setPendingQty(int pendingQty) {
+	public void setPendingQty(float pendingQty) {
 		this.pendingQty = pendingQty;
 	}
-	public int getIndedQty() {
+	public float getIndedQty() {
 		return indedQty;
 	}
-	public void setIndedQty(int indedQty) {
+	public void setIndedQty(float indedQty) {
 		this.indedQty = indedQty;
 	}
 	public float getDiscPer() {
@@ -203,10 +203,10 @@ public class PoDetail {
 		this.landingCost = landingCost;
 	}
 	
-	public int getBalanceQty() {
+	public float getBalanceQty() {
 		return balanceQty;
 	}
-	public void setBalanceQty(int balanceQty) {
+	public void setBalanceQty(float balanceQty) {
 		this.balanceQty = balanceQty;
 	}
 	@Override
