+++ src/main/java/com/ats/tril/model/indent/IndentTrans.java
@@ -13,7 +13,7 @@ public class IndentTrans {
 	
 	private int itemId;
 	
-	private int indQty;
+	private float indQty;
 	
 	private String indItemUom;
 	
@@ -29,7 +29,7 @@ public class IndentTrans {
 	
 	private int indDStatus;
 	
-	private int	indFyr;
+	private float	indFyr;
 	
 	private int delStatus;
 	
@@ -68,7 +68,7 @@ public class IndentTrans {
 		return itemId;
 	}
 
-	public int getIndQty() {
+	public float getIndQty() {
 		return indQty;
 	}
 
@@ -100,7 +100,7 @@ public class IndentTrans {
 		return indDStatus;
 	}
 
-	public int getIndFyr() {
+	public float getIndFyr() {
 		return indFyr;
 	}
 
@@ -121,7 +121,7 @@ public class IndentTrans {
 		this.itemId = itemId;
 	}
 
-	public void setIndQty(int indQty) {
+	public void setIndQty(float indQty) {
 		this.indQty = indQty;
 	}
 
@@ -153,7 +153,7 @@ public class IndentTrans {
 		this.indDStatus = indDStatus;
 	}
 
-	public void setIndFyr(int indFyr) {
+	public void setIndFyr(float indFyr) {
 		this.indFyr = indFyr;
 	}
